+++ src/main/java/com/nhom7/coworkingspace/constant/OtpPurpose.java
@@ -1,0 +1,6 @@
+package com.nhom7.coworkingspace.constant;
+
+public enum OtpPurpose {
+    ACCOUNT_CONFIRMATION,
+    PASSWORD_RESET
+}